#11208: Refactor ProgramCache to remove nested type erasure
